Move query params to query.go
